categorical indexes refactoring
